Refactor record form
